Improve and make test calculations more realistic.
To test the correctness of the calculations, the test route now consists
of five points taken with the MatLab mobile app on board of a real bus.
Also, the example fuel bus parameters are taken from a Mercedes-Benz
Citaro.

However, it was necesarry to add a step in the distances calculation to
convert the latitude and longitude differences to meters. Additionaly,
the angle calculation has been removed because it was only used to
calculates sines, wich can be easily done dividing the height difference
by the distance.