Start web application prototype
Start the prototype of the web application to interact with the
simulator using Streamlit.

The application allows uploading a route from a CSV file and visualizing
it on an interactive map, which displays the route, altitudes, speeds,
and accelerations. To achieve this, a classmethod was added to the
`Route` class to create a route from a dataframe.

It also provides fields to enter bus parameters and a button to run the
simulation, then displays the energy consumed and the remaining energy
percentage.