Add `simulate_trip()` method to the Bus class
This method performs a simulation of a bus trip along a route and
returns a dictionary with all kinds of simulation results, which will
then be displayed in the web interface.

The method has an optional argument `modify_bus`. If `True`, it will
modify the state of the bus according to the simulation results.
Otherwise, these results will be returned but the `Bus` object will
remain unchanged.

Additionally, a property `engine_type` has been added to the Bus class,
which returns the type of the bus's engine.