Add scripts for preprocessing KML and MatLab data
Two scripts have been added to facilitate preprocessing of route data so
that it can be easily imported into the application.

- The first script converts `.kml` files generated by Geo Tracker into
  `.csv` files that can be imported directly.
- The second script takes a `.csv` file generated when exporting data
  from the mobile version of MatLab using the `writetimetable` command.
  The script converts date strings to timestamps, sets the initial
  instant to zero, and ensures that the column names match what Sidrobus
  expects, then saves the result to a new `.csv` file.

The functionality of both scripts can also be imported from a new module
called `sidrobus.preprocessing`. Instead of passing files to the
functions, they receive either a string with the contents of the `.kml`
file or a pandas `DataFrame` with the MatLab data, and return a
preprocessed `DataFrame`. This way, they can be used directly in the
application code if desired, and will be useful for implementing
preprocessing functionality directly in the web application.