Add script to filter route altitudes
Add a script that allows filtering the altitudes of routes to remove
noise, with the option to choose between a Gaussian filter or a
Savitzky-Golay filter.

The script takes as input a CSV file with route data and saves the
result to another CSV file. Its functionality has also been included in
the `sidrobus.preprocessing` module so it can be used from Python, for
example, in the web application.